abir commit

+++ app/src/main/java/com/selsela/example/data/model/home/Product.java
@@ -396,7 +396,7 @@ public class Product extends SugarRecord implements Parcelable {
         }
     };
 
-    public double getProductId() {
+    public int getProductId() {
         return productId;
     }
